+++ packages/box/tools/build.js
@@ -1,3 +1,3 @@
-const build = require('../../../tools/build');
+const build = require('../../../tools/rollup');
 
 build(require('../package.json'));