+++ packages/libraries/styling/src/index.ts
@@ -5,12 +5,9 @@ import styled from 'styled-components';
  * styled-system (with css prop util) and styled-components
  */
 export * from './styled-components';
-export {
-  AnyStyledComponent,
+export type {
   BaseThemeProviderComponent,
   BaseWithThemeFnInterface,
-  createGlobalStyle,
-  css,
   CSSKeyframes,
   CSSObject,
   CSSProp,
@@ -18,13 +15,7 @@ export {
   CSSPseudos,
   DefaultTheme,
   GlobalStyleComponent,
-  isStyledComponent,
-  ThemeConsumer as PrimitiveThemeConsumer,
-  ThemeContext as PrimitiveThemeContext,
-  ThemeProvider as PrimitiveThemeProvider,
   ThemeProviderProps as PrimitiveThemeProviderProps,
-  useTheme as primitiveUseTheme,
-  withTheme as primitiveWithTheme,
   StyledComponent,
   StyledComponentBase,
   StyledComponentInnerAttrs,
@@ -47,6 +38,16 @@ export {
   ThemeProps,
   ThemeProviderComponent,
   WithThemeFnInterface,
+} from 'styled-components';
+export {
+  createGlobalStyle,
+  css,
+  isStyledComponent,
+  ThemeConsumer as PrimitiveThemeConsumer,
+  ThemeContext as PrimitiveThemeContext,
+  ThemeProvider as PrimitiveThemeProvider,
+  useTheme as primitiveUseTheme,
+  withTheme as primitiveWithTheme,
 } from 'styled-components';
 export * from 'styled-system';
 /**